🛠️ model.py -> Optimized data types for efficiency 🛠️ train.py -> Increased BATCH_SIZE from 4 to 8

+++ model.py
@@ -1,17 +1,19 @@
 import jax
 
 def feed_forward(params, x):
-    x = jax.numpy.dot(x, params['in_weight'])
+    x = x.astype(jax.numpy.float16)
+    x = jax.numpy.dot(x, params['in_weight'].astype(jax.numpy.float16))
     x = jax.nn.gelu(x, approximate=True)
-    x = jax.numpy.dot(x, params['out_weight'])
-    return x
+    x = jax.numpy.dot(x, params['out_weight'].astype(jax.numpy.float16))
+    return x.astype(jax.numpy.float32)
 
 def attention(params, x, mask, n_heads, scale):
     batch_size, seq_len, _ = x.shape
     # Linear transformations
-    q = jax.numpy.dot(x, params['q_linear']).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3)
-    k = jax.numpy.dot(x, params['k_linear']).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3)
-    v = jax.numpy.dot(x, params['v_linear']).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3)
+    x = x.astype(jax.numpy.float16)
+    q = jax.numpy.dot(x, params['q_linear'].astype(jax.numpy.float16)).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3).astype(jax.numpy.float32)
+    k = jax.numpy.dot(x, params['k_linear'].astype(jax.numpy.float16)).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3).astype(jax.numpy.float32)
+    v = jax.numpy.dot(x, params['v_linear'].astype(jax.numpy.float16)).reshape(batch_size, seq_len, n_heads, -1).transpose(0, 2, 1, 3).astype(jax.numpy.float32)
     # Compute attention scores
     scores = jax.numpy.matmul(q, k.transpose(0, 1, 3, 2)) * scale
     scores = jax.numpy.where(mask[:seq_len, :seq_len], -jax.numpy.inf, scores)
@@ -19,8 +21,8 @@ def attention(params, x, mask, n_heads, scale):
     # Compute output
     output = jax.numpy.matmul(scores, v)
     output = output.transpose(0, 2, 1, 3).reshape(batch_size, seq_len, -1)
-    output = jax.numpy.dot(output, params['o_linear'])
-    return output
+    output = jax.numpy.dot(output.astype(jax.numpy.float16), params['o_linear'].astype(jax.numpy.float16))
+    return output.astype(jax.numpy.float32)
 
 def transformer_block(params, x, mask, n_heads, scale):
     r = attention(params['attention'], layer_norm(x, params['attention_norm_scale'], params['attention_norm_bias']), mask, n_heads, scale)
@@ -43,39 +45,39 @@ def language_model(params, token_ids, pos, mask, n_heads, scale):
     x = layer_norm(x, params['out_norm_scale'], params['out_norm_bias'])
     return jax.numpy.dot(x, params['out_linear_weight'])
 
-def init_params(vocab_size, seq_len, num_blocks=16, num_heads=8, hidden_dim=768, ff_dim=2048, rng_key=jax.random.PRNGKey(0), dtype=jax.numpy.float32):
+def init_params(vocab_size, seq_len, num_blocks=16, num_heads=8, hidden_dim=768, ff_dim=2048, rng_key=jax.random.PRNGKey(0)):
     rng_key, subkey = jax.random.split(rng_key)
-    xavier_uniform_init = jax.nn.initializers.glorot_uniform(dtype=dtype)
-    kaiming_normal_init = jax.nn.initializers.he_normal(dtype=dtype)
+    xavier_uniform_init = jax.nn.initializers.glorot_uniform(dtype=jax.numpy.float32)
+    kaiming_normal_init = jax.nn.initializers.he_normal(dtype=jax.numpy.float32)
 
     learnable_params = {
-        'tok_emb': jax.random.normal(subkey, (vocab_size, hidden_dim), dtype=dtype) * 0.02,
-        'pos_emb': jax.random.normal(subkey, (vocab_size, hidden_dim), dtype=dtype) * 0.02,
-        'pos_norm_scale': jax.numpy.ones(hidden_dim, dtype=dtype),
-        'pos_norm_bias': jax.numpy.zeros(hidden_dim, dtype=dtype),
+        'tok_emb': jax.random.normal(subkey, (vocab_size, hidden_dim), dtype=jax.numpy.float32) * 0.02,
+        'pos_emb': jax.random.normal(subkey, (vocab_size, hidden_dim), dtype=jax.numpy.float32) * 0.02,
+        'pos_norm_scale': jax.numpy.ones(hidden_dim, dtype=jax.numpy.float32),
+        'pos_norm_bias': jax.numpy.zeros(hidden_dim, dtype=jax.numpy.float32),
         'transformer_blocks': [],
-        'out_norm_scale': jax.numpy.ones(hidden_dim, dtype=dtype),
-        'out_norm_bias': jax.numpy.zeros(hidden_dim, dtype=dtype),
-        'out_linear_weight': xavier_uniform_init(rng_key, (hidden_dim, vocab_size), dtype=dtype),
+        'out_norm_scale': jax.numpy.ones(hidden_dim, dtype=jax.numpy.float32),
+        'out_norm_bias': jax.numpy.zeros(hidden_dim, dtype=jax.numpy.float32),
+        'out_linear_weight': xavier_uniform_init(rng_key, (hidden_dim, vocab_size), dtype=jax.numpy.float32),
     }
 
     for _ in range(num_blocks):
         rng_key, block_key = jax.random.split(rng_key)
         block_params = {
             'attention': {
-                'q_linear': xavier_uniform_init(block_key, (hidden_dim, num_heads * (hidden_dim // num_heads)), dtype=dtype),
-                'k_linear': xavier_uniform_init(block_key, (hidden_dim, num_heads * (hidden_dim // num_heads)), dtype=dtype),
-                'v_linear': xavier_uniform_init(block_key, (hidden_dim, num_heads * (hidden_dim // num_heads)), dtype=dtype),
-                'o_linear': xavier_uniform_init(block_key, (num_heads * (hidden_dim // num_heads), hidden_dim), dtype=dtype),
+                'q_linear': xavier_uniform_init(block_key, (hidden_dim, num_heads * (hidden_dim // num_heads)), dtype=jax.numpy.float32),
+                'k_linear': xavier_uniform_init(block_key, (hidden_dim, num_heads * (hidden_dim // num_heads)), dtype=jax.numpy.float32),
+                'v_linear': xavier_uniform_init(block_key, (hidden_dim, num_heads * (hidden_dim // num_heads)), dtype=jax.numpy.float32),
+                'o_linear': xavier_uniform_init(block_key, (num_heads * (hidden_dim // num_heads), hidden_dim), dtype=jax.numpy.float32),
             },
             'feed_forward': {
-                'in_weight': kaiming_normal_init(block_key, (hidden_dim, ff_dim), dtype=dtype),
-                'out_weight': xavier_uniform_init(block_key, (ff_dim, hidden_dim), dtype=dtype),
+                'in_weight': kaiming_normal_init(block_key, (hidden_dim, ff_dim), dtype=jax.numpy.float32),
+                'out_weight': xavier_uniform_init(block_key, (ff_dim, hidden_dim), dtype=jax.numpy.float32),
             },
-            'attention_norm_scale': jax.numpy.ones(hidden_dim, dtype=dtype),
-            'attention_norm_bias': jax.numpy.zeros(hidden_dim, dtype=dtype),
-            'ffn_norm_scale': jax.numpy.ones(hidden_dim, dtype=dtype),
-            'ffn_norm_bias': jax.numpy.zeros(hidden_dim, dtype=dtype),
+            'attention_norm_scale': jax.numpy.ones(hidden_dim, dtype=jax.numpy.float32),
+            'attention_norm_bias': jax.numpy.zeros(hidden_dim, dtype=jax.numpy.float32),
+            'ffn_norm_scale': jax.numpy.ones(hidden_dim, dtype=jax.numpy.float32),
+            'ffn_norm_bias': jax.numpy.zeros(hidden_dim, dtype=jax.numpy.float32),
         }
         learnable_params['transformer_blocks'].append(block_params)
